requirements.txt
wheel==0.42.0
transformers==4.44.2
accelerate==0.27.2
opencv-python==4.8.0.76
diffusers==0.30.3
numpy==1.26.4
nvgpu==0.10.0
omegaconf==2.3.0
ninja==1.11.1.1
xformers==0.0.23.post1
pydantic==2.6.1
torchsde==0.2.6
SpeechRecognition==3.10.4
googletrans==4.0.0-rc1
scipy==1.12.0
pydub==0.25.1
pyyaml==6.0.1
librosa==0.9.2
peft==0.12.0
sentencepiece==0.2.0
ffmpeg-python==0.2.0